Don't take into account fake alpha when finding image bounding box
When calculating the bounding box of the non-magenta area of a
sprite (as part of encoding a PIL Image object to a sprite),
don't use the alpha channel.

If an image has no civ colors, i.e. the alpha channel is all zero,
the RGBA image would be transparent and the whole bounding box
empty. But the alpha channel isn't for transparency, so ignore it.