Drop unimplemented --storyboard and --adjacent-mask flags
I will not continue development of the XML output format so I'm
not going to implement those there.

For the text output format on the other hand, I've *only* implemented
these options, and I'm not going to implement the alternatives.

Having images together in "storyboard" images with masks is clearly
superior to a sea of individual images and masks.

The most elaborate animation from the originals in the game has 41
frames. Even that is still manageable in a single image.

The SPR format can re-use images in multiple animations, but this
happens only once in the game originals, in Fantasy/Unit79.spr, and
there it's a bug. My code handles this gracefully anyway, by storing
such images only in the first animation storyboard that use them.

I do still have one TODO: some of the existing SPR files contain
images that are never actually used in any animations. Now that I
create a storyboard only from the images that are used, these are
never exported.

This is not a problem per se, except that I want users to be able
to access all available game art. So I still need to export those
separately somehow.