Require image dir to not exist and don't leave behind half-written files
Before this change, if the .txt/.json output file already existed,
but the matching images directory did not, we'd succeed in writing
the images, but then fail to write the config file, leaving us with
half a conversion.

Make sure that doesn't happen.

Open the .txt/.json file first, so we know both the config and the
dir could be created before we write anything.

Also make sure that, if the config file was successfully created,
but then creation of the dir fails, the config file (which did not
exist before and has not been written to) is removed again.

Don't add any option to write to existing files/dirs to avoid
all the potential complications that can come with that.