Drop unused --with-border output flag for now
I haven't set up any plumbing for this, and I haven't figured out
yet how to pass arguments to the load/save functions in a way that
I like.

If I'm gonna have input/output arguments, I want to be able to pass
only the arguments relevant for the file format that's being loaded
or saved.